refactor: extract simulation controls panel into dedicated component
- Create SimulationControlsPanel component in components/ui/
- Extract ~256 lines of control rendering logic from app/page.tsx
- Preserve special handling for cellular-automaton and four-pole-gradient patterns
- Maintain responsive grid behavior based on sidebar width
- Add comprehensive AIDEV-NOTE documentation
- Improve code maintainability and prepare for future UI features

Resolves #17
